// src/enums/error-message.ts
export enum ErrorMessage {
  BadRequest = 'Запрос некорректный или неверный.',
  NotFound = 'Запрошенный ресурс не был найден на сервере.',
  InternalServerError = 'На сервере произошла внутренняя ошибка.'
}